+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/dialog/index.ts
@@ -2,7 +2,7 @@ export * from './apply-configuration/o-table-apply-configuration-dialog.componen
 export * from './export/o-table-export-dialog.component';
 export * from './filter-by-column/o-table-filter-by-column-data-dialog.component';
 export * from './load-filter/o-table-load-filter-dialog.component';
-// export * from './o-table-base-dialog.class';
 export * from './store-configuration/o-table-store-configuration-dialog.component';
 export * from './store-filter/o-table-store-filter-dialog.component';
 export * from './visible-columns/o-table-visible-columns-dialog.component';
+export * from './group-by-columns/o-table-group-by-columns-dialog.component';