+++ projects/ontimize-web-ngx/src/lib/components/table/o-table.component.ts
@@ -1,8 +1,12 @@
+import { animate, state, style, transition, trigger } from '@angular/animations';
 import { SelectionChange, SelectionModel } from '@angular/cdk/collections';
+import { DomPortalOutlet, TemplatePortal } from '@angular/cdk/portal';
 import {
   AfterViewInit,
+  ApplicationRef,
   ChangeDetectionStrategy,
   Component,
+  ComponentFactoryResolver,
   ContentChild,
   ContentChildren,
   ElementRef,
@@ -18,14 +22,12 @@ import {
   TemplateRef,
   ViewChild,
   ViewChildren,
-  ViewEncapsulation,
-  ViewRef,
   ViewContainerRef,
-  ApplicationRef,
-  ComponentFactoryResolver
+  ViewEncapsulation,
+  ViewRef
 } from '@angular/core';
 import { MatCheckboxChange, MatDialog, MatMenu, MatPaginator, MatTab, MatTabGroup, PageEvent } from '@angular/material';
-import { BehaviorSubject, combineLatest, Observable, of, Subscription } from 'rxjs';
+import { BehaviorSubject, combineLatest, forkJoin, Observable, of, Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 import { BooleanConverter, InputConverter } from '../../decorators/input-converter';
@@ -35,11 +37,13 @@ import { OTableButton } from '../../interfaces/o-table-button.interface';
 import { OTableButtons } from '../../interfaces/o-table-buttons.interface';
 import { OTableDataSource } from '../../interfaces/o-table-datasource.interface';
 import { OTableMenu } from '../../interfaces/o-table-menu.interface';
+import { OnClickTableEvent } from '../../interfaces/o-table-onclick.interface';
 import { OTableOptions } from '../../interfaces/o-table-options.interface';
 import { OTablePaginator } from '../../interfaces/o-table-paginator.interface';
 import { OTableQuickfilter } from '../../interfaces/o-table-quickfilter.interface';
 import { OntimizeServiceProvider } from '../../services/factories';
 import { SnackBarService } from '../../services/snackbar.service';
+import { TableFilterByColumnDialogResult } from '../../types';
 import { Expression } from '../../types/expression.type';
 import { OColumnAggregate } from '../../types/o-column-aggregate.type';
 import { ColumnValueFilterOperator, OColumnValueFilter } from '../../types/o-column-value-filter.type';
@@ -47,6 +51,7 @@ import { OPermissions } from '../../types/o-permissions.type';
 import { OTableInitializationOptions } from '../../types/o-table-initialization-options.type';
 import { OTableMenuPermissions } from '../../types/o-table-menu-permissions.type';
 import { OTablePermissions } from '../../types/o-table-permissions.type';
+import { OTableGroupedRow } from '../../types/o-table-row-group.type';
 import { OQueryDataArgs } from '../../types/query-data-args.type';
 import { QuickFilterFunction } from '../../types/quick-filter-function.type';
 import { SQLOrder } from '../../types/sql-order.type';
@@ -60,27 +65,23 @@ import { Util } from '../../util/util';
 import { OContextMenuComponent } from '../contextmenu/o-context-menu.component';
 import { OFormComponent } from '../form/o-form.component';
 import { DEFAULT_INPUTS_O_SERVICE_COMPONENT, OServiceComponent } from '../o-service-component.class';
+import { OBaseTableCellRenderer, OTableCellRendererServiceComponent } from './column';
 import { OTableColumnCalculatedComponent } from './column/calculated/o-table-column-calculated.component';
 import { OColumn } from './column/o-column.class';
 import { OTableColumnComponent } from './column/o-table-column.component';
 import { DefaultOTableOptions } from './extensions/default-o-table-options.class';
-import {
-  OTableFilterByColumnDataDialogComponent
-} from './extensions/dialog/filter-by-column/o-table-filter-by-column-data-dialog.component';
+import { OTableFilterByColumnDataDialogComponent } from './extensions/dialog/filter-by-column/o-table-filter-by-column-data-dialog.component';
 import { OBaseTablePaginator } from './extensions/footer/paginator/o-base-table-paginator.class';
+import { OFilterColumn } from './extensions/header/table-columns-filter/columns/o-table-columns-filter-column.component';
 import { OTableColumnsFilterComponent } from './extensions/header/table-columns-filter/o-table-columns-filter.component';
 import { OTableInsertableRowComponent } from './extensions/header/table-insertable-row/o-table-insertable-row.component';
 import { OTableOptionComponent } from './extensions/header/table-option/o-table-option.component';
 import { OTableDataSourceService } from './extensions/o-table-datasource.service';
 import { OTableStorage } from './extensions/o-table-storage.class';
 import { OTableDao } from './extensions/o-table.dao';
+import { OTableRowExpandableComponent, OTableRowExpandedChange } from './extensions/row/table-row-expandable/o-table-row-expandable.component';
 import { OMatSort } from './extensions/sort/o-mat-sort';
 import { OMatSortHeader } from './extensions/sort/o-mat-sort-header';
-import { OFilterColumn } from './extensions/header/table-columns-filter/columns/o-table-columns-filter-column.component';
-import { trigger, state, transition, style, animate } from '@angular/animations';
-import { OTableRowExpandableComponent, OTableRowExpandedChange } from './extensions/row/table-row-expandable/o-table-row-expandable.component';
-import { TemplatePortal, DomPortalOutlet } from '@angular/cdk/portal';
-import { OnClickTableEvent } from '../../interfaces/o-table-onclick.interface';
 
 export const DEFAULT_INPUTS_O_TABLE = [
   ...DEFAULT_INPUTS_O_SERVICE_COMPONENT,
@@ -176,7 +177,16 @@ export const DEFAULT_INPUTS_O_TABLE = [
   'rowClass: row-class',
 
   // filter-column-active-by-default [yes|no|true|false]: show icon filter by default in the table
-  'filterColumnActiveByDefault:filter-column-active-by-default'
+  'filterColumnActiveByDefault:filter-column-active-by-default',
+
+  //grouped-columns [string]: grouped columns separated by ';'. Default: no value.
+  'groupedColumns: grouped-columns',
+
+  //groupable[boolean]: Indicates whether or not the column can be groupable. By default: false
+  'groupable',
+
+  //collapse-grouped-columns [yes|no|true|false]: Whether collapse the grouped columns by default
+  'collapseGroupedColumns: collapse-grouped-columns'
 ];
 
 export const DEFAULT_OUTPUTS_O_TABLE = [
@@ -239,6 +249,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
 
   _filterColumns: Array<OFilterColumn>;
   portalHost: DomPortalOutlet;
+  onDataLoadedCellRendererSubscription: Subscription;
 
   get diameterSpinner() {
     const minHeight = OTableComponent.DEFAULT_BASE_SIZE_SPINNER;
@@ -351,6 +362,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   resizable: boolean = true;
   @InputConverter()
   autoAdjust: boolean = false;
+  @InputConverter()
+  groupable: boolean = false;
+  @InputConverter()
+  collapseGroupedColumns: boolean = false;
 
   protected _enabled: boolean = true;
   get enabled(): boolean {
@@ -380,6 +395,8 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   public daoTable: OTableDao | null;
   public dataSource: OTableDataSource | null;
   public visibleColumns: string;
+  public groupedColumns: string;
+
   public sortColumns: string;
   public rowClass: (rowData: any, rowIndex: number) => string | string[];
 
@@ -528,6 +545,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   @ViewChild('exportOptsTemplate', { static: false })
   exportOptsTemplate: TemplateRef<any>;
 
+  public groupedColumnsArray: string[] = [];
   @HostListener('window:resize', [])
   updateScrolledState(): void {
     if (this.horizontalScroll) {
@@ -574,6 +592,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     if (this.oTableButtons) {
       this.oTableButtons.registerButtons(this.tableButtons.toArray());
     }
+
   }
 
   ngAfterViewInit() {
@@ -680,6 +699,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
+  /**
+   * Allow reinitialize table adding options
+   * @param options
+   */
   reinitialize(options: OTableInitializationOptions): void {
     if (options) {
       const clonedOpts = Object.assign({}, options);
@@ -725,6 +748,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     this.parseVisibleColumns();
     this.setDatasource();
     this.registerDataSourceListeners();
+    this.parseGroupedColumns();
     this.parseSortColumns();
     this.registerSortListener();
     this.setFiltersConfiguration(this.state);
@@ -759,6 +783,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
         this.asyncLoadSubscriptions[idx].unsubscribe();
       }
     });
+
   }
 
   /**
@@ -1161,6 +1186,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return this.expandableItem.selected;
   }
 
+  /**
+   * Allow to expand or collapse the expandable row.
+   * @param item
+   * @param rowIndex
+   * @param event
+   */
   public toogleRowExpandable(item: any, rowIndex: number, event: Event): void {
     event.stopPropagation();
     event.preventDefault();
@@ -1425,6 +1456,9 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     //
   }
 
+  /**
+   * Triggers navigation to new item insertion
+   */
   add() {
     if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_INSERT)) {
       return;
@@ -1432,6 +1466,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     super.insertDetail();
   }
 
+  /**
+   * Removes selected rows
+   * @param [clearSelectedItems]
+   */
   remove(clearSelectedItems: boolean = false) {
     if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_DELETE)) {
       return;
@@ -1459,10 +1497,16 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
+  /**
+   * Refreshs table component
+   */
   refresh() {
     this.reloadData();
   }
 
+  /**
+   * Shows and select all checkbox
+   */
   showAndSelectAllCheckbox() {
     if (this.isSelectionModeMultiple()) {
       if (this.selectAllCheckbox) {
@@ -1481,6 +1525,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
+  /**
+   * Reloads data
+   * @returns
+   */
   reloadData() {
     if (!this.checkEnabledActionPermission(PermissionsUtils.ACTION_REFRESH)) {
       return;
@@ -1750,11 +1798,16 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
       }
 
       let itemId: string = '';
-      const keysLenght = self.keysArray.length;
-      self.keysArray.forEach((key: string, idx: number) => {
-        const suffix = idx < (keysLenght - 1) ? ';' : '';
-        itemId += item[key] + suffix;
-      });
+
+      if (this.isGroup(index, item)) {
+        itemId += item.column;
+      } else {
+        const keysLenght = self.keysArray.length;
+        self.keysArray.forEach((key: string, idx: number) => {
+          const suffix = idx < (keysLenght - 1) ? ';' : '';
+          itemId += item[key] + suffix;
+        });
+      }
 
 
       const asyncAndVisible = self.asyncLoadColumns.filter(c => self._oTableOptions.visibleColumns.indexOf(c) !== -1);
@@ -1801,22 +1854,47 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
+
+  /**
+   * Returns the current page table data.
+   * @returns
+   */
   getValue() {
     return this.dataSource.getCurrentData();
   }
 
+
+  /**
+   * Gets all values
+   * @returns
+   */
   getAllValues() {
     return this.dataSource.getCurrentAllData();
   }
 
+
+  /**
+   * Gets all rendered values
+   * @returns
+   */
   getAllRenderedValues() {
     return this.dataSource.getAllRendererData();
   }
 
+
+  /**
+   * Returns the current page table renderer data.
+   * @returns
+   */
   getRenderedValue() {
     return this.dataSource.getCurrentRendererData();
   }
 
+
+  /**
+   * Gets sql types from data source
+   * @returns
+   */
   getSqlTypes() {
     return Util.isDefined(this.dataSource.sqlTypes) ? this.dataSource.sqlTypes : {};
   }
@@ -1847,8 +1925,16 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return sortColumnsFilter;
   }
 
+  get originalGroupedColumnsArray(): Array<string> {
+    return Util.parseArray(this.groupedColumns, true);
+  }
+
   getStoredColumnsFilters() {
     return this.oTableStorage.getStoredColumnsFilters();
+  }
+
+  getStoredGroupedColumns() {
+    return this.oTableStorage.getStoredGroupedColumns();
   }
 
   onFilterByColumnClicked() {
@@ -1875,18 +1961,36 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     }
   }
 
+  /**
+   * Clear all filters(column filter and quickfilter) and reload data
+   * @param [triggerDatasourceUpdate]
+   */
   clearFilters(triggerDatasourceUpdate: boolean = true): void {
     this.dataSource.clearColumnFilters(triggerDatasourceUpdate);
     if (this.oTableMenu && this.oTableMenu.columnFilterOption) {
       this.oTableMenu.columnFilterOption.setActive(this.showFilterByColumnIcon);
     }
+    this.onFilterByColumnChange.emit(this.dataSource.getColumnValueFilters());
     if (this.oTableQuickFilterComponent) {
       this.oTableQuickFilterComponent.setValue(void 0);
     }
   }
 
+  clearColumnFilter(attr: string, triggerDatasourceUpdate: boolean = true): void {
+    this.dataSource.clearColumnFilter(attr, triggerDatasourceUpdate);
+    this.onFilterByColumnChange.emit(this.dataSource.getColumnValueFilters());
+    this.reloadPaginatedDataFromStart();
+  }
+
+  filterByColumn(columnValueFilter: OColumnValueFilter) {
+    this.dataSource.addColumnFilter(columnValueFilter);
+    this.onFilterByColumnChange.emit(this.dataSource.getColumnValueFilters());
+    this.reloadPaginatedDataFromStart();
+  }
+
   clearColumnFilters(triggerDatasourceUpdate: boolean = true): void {
     this.dataSource.clearColumnFilters(triggerDatasourceUpdate);
+    this.onFilterByColumnChange.emit(this.dataSource.getColumnValueFilters());
     this.reloadPaginatedDataFromStart();
   }
 
@@ -1916,19 +2020,26 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
         preloadValues: this.oTableColumnsFilterComponent.preloadValues,
         mode: this.oTableColumnsFilterComponent.mode
       },
+      minWidth: '380px',
       disableClose: true,
       panelClass: ['o-dialog-class', 'o-table-dialog']
     });
+
     dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        const columnValueFilter = dialogRef.componentInstance.getColumnValuesFilter();
-        //guardar en localstorage el cambio
-        const sortedFilterableColumn = dialogRef.componentInstance.getFilterColumn();
-        this.storeFilterColumns(sortedFilterableColumn);
-        this.dataSource.addColumnFilter(columnValueFilter);
-        this.onFilterByColumnChange.emit(this.dataSource.getColumnValueFilters());
-        this.reloadPaginatedDataFromStart();
+      switch (result) {
+        case TableFilterByColumnDialogResult.ACCEPT:
+          const columnValueFilter = dialogRef.componentInstance.getColumnValuesFilter();
+          this.filterByColumn(columnValueFilter);
+          break;
+        case TableFilterByColumnDialogResult.CLEAR:
+          const col = dialogRef.componentInstance.column;
+          this.clearColumnFilter(col.attr);
+          break;
       }
+    });
+    dialogRef.componentInstance.onSortFilterValuesChange.subscribe(sortedFilterableColumn => {
+      // guardar en localstorage el cambio
+      this.storeFilterColumns(sortedFilterableColumn);
     });
   }
 
@@ -2141,10 +2252,20 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return this.daoTable.updateQuery(filter, updateData, sqlTypesArg);
   }
 
+
+  /**
+   * Gets data table
+   * @returns
+   */
   getDataArray() {
     return this.daoTable.data;
   }
 
+
+  /**
+   * Sets data table
+   * @param data
+   */
   setDataArray(data: Array<any>) {
     if (this.daoTable) {
       // remote pagination has no sense when using static-data
@@ -2280,6 +2401,10 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
           case 'columns-filter':
             this.setFiltersConfiguration(conf);
             break;
+          case 'grouped-columns':
+            this.state['grouped-columns'] = conf['grouped-columns']
+            this.parseGroupedColumns();
+            break;
           case 'page':
             this.state.currentPage = conf.currentPage;
             this.currentPage = conf.currentPage;
@@ -2378,6 +2503,7 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     colDef.searchable = false;
     colDef.orderable = false;
     colDef.resizable = false;
+    colDef.groupable = false;
     colDef.title = undefined;
     colDef.width = '48px';
     this.pushOColumnDefinition(colDef);
@@ -2408,6 +2534,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     Util.copyToClipboard(JSON.stringify(selectedItems));
   }
 
+
+  /**
+   * Triggers navigation to item detail, receiving item data
+   * @param item
+   * @returns detail
+   */
   viewDetail(item: any): void {
     if (!this.checkEnabledActionPermission('detail')) {
       return;
@@ -2415,6 +2547,12 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     super.viewDetail(item);
   }
 
+
+  /**
+   * Triggers navigation to item edition, receiving item data
+   * @param item
+   * @returns detail
+   */
   editDetail(item: any): void {
     if (!this.checkEnabledActionPermission('edit')) {
       return;
@@ -2485,7 +2623,8 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     if (table) {
       instance.setDefaultProperties({
         orderable: this.orderable,
-        resizable: this.resizable
+        resizable: this.resizable,
+        groupable: this.groupable
       });
     }
     if (column) {
@@ -2505,4 +2644,140 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return col.DOMWidth || this.getThWidthFromOColumn(col);
   }
 
+  public getMinWidthColumn(col: OColumn): string {
+    return Util.extractPixelsValue(col.minWidth, Codes.DEFAULT_COLUMN_MIN_WIDTH) + 'px';
+  }
+
+  showExpandedColumn(): boolean {
+    return Util.isDefined(this.tableRowExpandable);
+  }
+
+  /**
+   * Parses grouped columns
+   */
+  parseGroupedColumns() {
+
+    this.groupedColumnsArray = this.state['grouped-columns'] || this.originalGroupedColumnsArray;
+    if (this.state['grouped-columns'] && this.state['initial-configuration']['grouped-columns']) {
+
+      const initialGroupedColumnsArray = this.state['initial-configuration']['grouped-columns'];
+
+      const difference = initialGroupedColumnsArray.filter(x => !this.originalGroupedColumnsArray.includes(x));
+      if (difference.length > 0) {
+        this.groupedColumnsArray = this.originalGroupedColumnsArray;
+      }
+    }
+
+    this.dataSource.updateGroupedColumns(this.groupedColumnsArray);
+  }
+
+  /**
+   * Groups by column
+   * @param column
+   */
+  public groupByColumn(column: OColumn) {
+    this.checkGroupByColumn(column.attr, true);
+    this.dataSource.updateGroupedColumns(this.groupedColumnsArray);
+  }
+
+  /**
+   * Ungroup by column
+   * @param column
+   */
+  unGroupByColumn(column: OColumn) {
+    this.checkGroupByColumn(column.attr, false);
+    this.dataSource.updateGroupedColumns(this.groupedColumnsArray);
+  }
+
+  /**
+   * Ungroup by all columns
+   */
+  unGroupByAllColumns() {
+    this.groupedColumnsArray = [];
+    this.dataSource.updateGroupedColumns(this.groupedColumnsArray);
+  }
+
+  checkGroupByColumn(field: string, add: boolean) {
+    let found = null;
+    for (const column of this.groupedColumnsArray) {
+      if (column === field) {
+        found = this.groupedColumnsArray.indexOf(column, 0);
+      }
+    }
+    if (found != null && found >= 0) {
+      if (!add) {
+        this.groupedColumnsArray.splice(found, 1);
+      }
+    } else {
+      if (add) {
+        this.groupedColumnsArray.push(field);
+      }
+    }
+  }
+
+  /**
+   * Determines whether item is group
+   * @param index
+   * @param item
+   * @returns true if group
+   */
+  isGroup(index, item): boolean {
+    return Util.isDefined(item.level);
+  }
+
+  /**
+   * Determines whether item is not group
+   * @param index
+   * @param item
+   * @returns true if not group
+   */
+  isNotGroup(index, item): boolean {
+    return !Util.isDefined(item.level);
+  }
+
+  groupHeaderClick(row: OTableGroupedRow) {
+    this.dataSource.toggleGroupByColumn(row);
+  }
+
+  getTextGroupRow(group: OTableGroupedRow) {
+    const field = this.groupedColumnsArray[group.level - 1];
+    let value = group.column[this.groupedColumnsArray[group.level - 1]];
+    const totalCounts = group.totalCounts;
+    const oCol = this.getOColumn(field);
+
+    if (!value && this.isInstanceOfOTableCellRendererServiceComponent(oCol.renderer)) {
+      value = ' - ';
+      if (!this.onDataLoadedCellRendererSubscription) {
+        this.onDataLoadedCellRendererSubscription = (oCol.renderer as any).onDataLoaded.subscribe(x => {
+          this.dataSource.updateGroupedColumns(this.groupedColumnsArray);
+        });
+      }
+    }
+    return this.translateService.get(oCol.title) + ': ' + value + ' (' + totalCounts + ')';
+
+  }
+
+  private isInstanceOfOTableCellRendererServiceComponent(renderer: OBaseTableCellRenderer) {
+    return (renderer as any).onDataLoaded;
+  }
+
+  /**
+   * Gets column data by attr
+   * @param attr
+   * @param row
+   * @returns column data by attr
+   */
+  getColumnDataByAttr(attr, row: any): any {
+    const oCol = this.getOColumn(attr);
+    const useRenderer = oCol && oCol.renderer && oCol.renderer.getCellData;
+    return useRenderer ? oCol.renderer.getCellData(row[oCol.attr], row) : row[oCol.attr];
+  }
+
+  getClassNameGroupHeader(row: OTableGroupedRow): string {
+    let className = '';
+    if (row.level <= 10) {
+      className = 'o-table-group-row o-table-group-row-level-' + row.level;
+    }
+    return className;
+  }
 }