+++ projects/ontimize-web-ngx/src/lib/i18n/i18n.ts
@@ -246,7 +246,19 @@ export const MAP = {
     'DATERANGE.last_month': 'Last month',
     'DATERANGE.this_year': 'This year',
     'DATERANGE.APPLYLABEL': 'Select dates',
-    'DATERANGE.W': 'W'
+    'DATERANGE.W': 'W',
+    'TABLE_CONTEXT_MENU.GROUP_BY': 'Group by',
+    'TABLE_CONTEXT_MENU.GROUP_BY_COLUMN': 'Group by {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_BY_COLUMN': 'UnGroup by {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_ALLCOLUMNS': 'UnGroup all',
+    'TABLE.BUTTONS.GROUPBY': 'Group / ungroup by',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.DESCRIPTION': 'Drag and drop item from the left (Available columns) to the right (Grouped columns). Reorder the items in Grouped columns as needed.',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.ALL_COLUMNS': 'Available columns',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.GROUPED_COLUMNS': 'Grouped columns',
+    'DUAL_LIST_SELECTOR.DROP_EMPTY_MESSAGE': 'Drag here to set the grouped columns',
+    'TABLE.VISIBLE_COLUMNS_DIALOG.ALL_COLUMNS': 'Available columns',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART1': 'Click on ',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART2': 'to hide/display the column. Use drag and drop to modify the column order.'
   },
 
   // Spanish
@@ -487,7 +499,21 @@ export const MAP = {
     'TABLE_CONTEXT_MENU.DELETE': 'Borrar',
     'TABLE_CONTEXT_MENU.FILTER': 'Filtrar',
     'TABLE_CONTEXT_MENU.FILTER_BY': 'Filter by:',
-    'TABLE_CONTEXT_MENU.FILTER_BY_VALUE': 'Filtrar por este valor'
+    'TABLE_CONTEXT_MENU.FILTER_BY_VALUE': 'Filtrar por este valor',
+    'TABLE_CONTEXT_MENU.GROUP_BY': 'Agrupar por',
+    'TABLE_CONTEXT_MENU.GROUP_BY_COLUMN': 'Agrupar por {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_BY_COLUMN': 'Desagrupar por {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_ALLCOLUMNS': 'Desagrupar todas las columnas',
+    'TABLE.BUTTONS.GROUPBY': 'Agrupar por',
+    'TABLE.DIALOG.TITLE_GROUPBY': 'Agrupar / desagrupar por columnas',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.DESCRIPTION': 'Arrastre y suelte de la izquierda (Columnas disponibles) a la derecha (Columnas agrupadas). Reordena las columnas agrupadas si fuera necesario',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.ALL_COLUMNS': 'Columnas disponibles',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.GROUPED_COLUMNS': 'Columnas agrupadas',
+    'DUAL_LIST_SELECTOR.DROP_EMPTY_MESSAGE': 'Arrastre aquí para establecer las columnas agrupadas',
+    'TABLE.VISIBLE_COLUMNS_DIALOG.ALL_COLUMNS': 'Columnas disponibles',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART1': 'Haz click sobre ',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART2': 'para ocultar/mostrar la columna. Utilice arrastrar y soltar para modificar el orden de las columnas.'
+
 
   },
   // Portuguese - BRAZIL
@@ -726,7 +752,18 @@ export const MAP = {
     'TABLE_CONTEXT_MENU.DELETE': 'Apagar',
     'TABLE_CONTEXT_MENU.FILTER': 'Filtrar',
     'TABLE_CONTEXT_MENU.FILTER_BY': 'Filtrar por:',
-    'TABLE_CONTEXT_MENU.FILTER_BY_VALUE': 'Filtrar por esse valor'
-
+    'TABLE_CONTEXT_MENU.FILTER_BY_VALUE': 'Filtrar por ese valor',
+    'TABLE_CONTEXT_MENU.GROUP_BY_COLUMN': 'Agrupar por {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_BY_COLUMN': 'Desagrupar por {0}',
+    'TABLE_CONTEXT_MENU.UNGROUP_ALLCOLUMNS': 'Desagrupar todas las columnas',
+    'TABLE.BUTTONS.GROUPBY': 'Agrupar / desagrupar por',
+    'TABLE.DIALOG.TITLE_GROUPBY': 'Agrupar / desagrupar por columnas',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.DESCRIPTION': 'Arrastre y suelte de la izquierda (Columnas disponibles) a la derecha (Columnas agrupadas). Reordena las columnas agrupadas si fuera necesario',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.AVAILABLE_COLUMNS': 'Columnas disponibles',
+    'TABLE_GROUP_BY_COLUMNS_DIALOG.GROUPED_COLUMNS': 'Columnas agrupadas',
+    'DUAL_LIST_SELECTOR.DROP_EMPTY_MESSAGE': 'Arrastre aquí para establecer las columnas agrupadas',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART1': 'Haz click sobre ',
+    'TABLE.VISIBLE_COLUMNS.DESCRIPTION_PART2': 'para ocultar/mostrar la columna. Utilice arrastrar y soltar para modificar el orden de las columnas.',
+    'TABLE.VISIBLE_COLUMNS_DIALOG.ALL_COLUMNS': 'Columnas disponibles'
   }
 };